+++ packages/app/src/components/Sliders/ParametricEditors/AngleEditor.tsx
@@ -6,6 +6,7 @@ import { useCompactMode } from '@/contexts/CompactModeContext'
 import { useKeyframeTarget } from '@/contexts/KeyframeTargetContext'
 import { useTimeline } from '@/contexts/TimelineContext'
 import { createDragHandler } from '@/utils/createDragHandler'
+import { keyframeEditedParam } from '@/utils/keyframeOnChange'
 import { scrollIntoViewAndFocusOnChange } from '@/utils/scrollIntoViewOnChange'
 import ui from './AngleEditor.module.css'
 import { useParamMetaCapture } from './paramMetaCapture'
@@ -107,15 +108,7 @@ export function AngleEditor(props: AngleEditorProps) {
           newAngle += 2 * Math.PI
         }
         props.setValue(newAngle)
-        // Auto-keyframe: only for already-animated params when auto mode is on
-        if (
-          timeline &&
-          props.dataParameterPath &&
-          timeline.autoKeyframe() &&
-          timeline.hasAnyKeyframes(props.dataParameterPath)
-        ) {
-          timeline.addKeyframeAtCurrentFrame(props.dataParameterPath)
-        }
+        keyframeEditedParam(timeline, props.dataParameterPath)
       }
       setAngle(initEvent)
       return {